Use a wait condition for joining threads
This seems to be more reliable than pthread_join(), which did not play
well with the thread cleanup on Mac OS X. This was reported as a crash
in <http://sourceforge.net/p/scm-pepper/bugs/7/>.

The new implementation is inspired by Qt's thread implementation for
unix. Thread::wait() waits for a wait condition and will be woken up
from within the cleanup function.