Don't use svn_sort__hash()
svn_sort__hash() is an internal function (despite being advertised in the public
documentation) that has been removed from public headers in svn-1.9. The
calls have been replaced by manual extraction and sorting of hash keys.

This ensures compatibility with svn-1.9. A new test case has been added that
verifies that the output of the `tree` call is sorted.